Clean up old directory enumeration code
