+++ bst/src/file_systems/fat/mod.rs
@@ -17,7 +17,6 @@
 mod bios_parameters_blocks;
 mod boot_sectors;
 mod directory_entries;
-mod directory_handles;
 mod fat_entries;
 mod file_system_info;
 mod reads;
